+++ static/scripts/pwd.js
@@ -1,14 +1,17 @@
 $(document).ready(() => {
 	$('.show_hide_password button').on('click', (event) => {
 		event.preventDefault();
-		if ($('.show_hide_password input').attr('type') == 'text') {
-			$('.show_hide_password input').attr('type', 'password');
-			$('.show_hide_password i').addClass('fa-eye-slash');
-			$('.show_hide_password i').removeClass('fa-eye');
-		} else if ($('.show_hide_password input').attr('type') == 'password') {
-			$('.show_hide_password input').attr('type', 'text');
-			$('.show_hide_password i').removeClass('fa-eye-slash');
-			$('.show_hide_password i').addClass('fa-eye');
+		const element = ($(event.target).is(':parent')) ? $(event.target).find('i') : $(event.target);
+		const passwordId = $($(event.target).parents('div')[1]).find('input').attr('id');
+
+		if ($('#' + passwordId).attr('type') == 'text') {
+			$('#' + passwordId).attr('type', 'password');
+			$(element).addClass('fa-eye-slash');
+			$(element).removeClass('fa-eye');
+		} else if ($('#' + passwordId).attr('type') == 'password') {
+			$('#' + passwordId).attr('type', 'text');
+			$(element).removeClass('fa-eye-slash');
+			$(element).addClass('fa-eye');
 		}
 	});
 });